Fixed serialization/deserialization small issue. Changed the amount of money from double to long, and money is now represented in cents, although this probably could be changed again if needed as it wasn't the main issue. Still a problem when: user 1 sends money to user 2 and user 2 sends back the money to user 1

+++ offlineeuro/src/main/java/nl/tudelft/trustchain/offlineeuro/communication/ICommunicationProtocol.kt
@@ -29,7 +29,7 @@ interface ICommunicationProtocol {
         publicKey: Element,
         bankName: String,
         challenge: BigInteger,
-        amount: Double
+        amount: Long
     ): BigInteger
 
     fun requestTransactionRandomness(
